Moved functions to cache.py, added response object
