+++ src/csvcubed/utils/uri.py
@@ -13,8 +13,6 @@ from urllib.parse import urlparse
 import rdflib
 from unidecode import unidecode
 
-from csvcubed.definitions import APP_ROOT_DIR_PATH
-
 _logger = logging.getLogger(__name__)
 
 _multiple_non_word_chars_regex = re.compile(r"[^\w]+")
@@ -116,74 +114,3 @@ def ensure_values_in_lists_looks_like_uris(values: list[str]) -> None:
             raise ValueError(f"'{value}' does not look like a URI.")
 
     _logger.debug("Values %s all look like URIs.", values)
-
-
-def get_url_to_file_path_map() -> Dict[str, Path]:
-    """
-    todo: add comment here
-    """
-
-    map_uri_to_file_path = {
-        "//purl.org/csv-cubed/qube-config/v1.0": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_0"
-        / "schema.json",
-        "//purl.org/csv-cubed/qube-config/v1.1": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_1"
-        / "schema.json",
-        "//purl.org/csv-cubed/qube-config/v1.2": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_2"
-        / "schema.json",
-        "//purl.org/csv-cubed/qube-config/v1.3": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_3"
-        / "schema.json",
-        "//purl.org/csv-cubed/qube-config/v1.4": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_4"
-        / "schema.json",
-        "//purl.org/csv-cubed/qube-config/v1": APP_ROOT_DIR_PATH
-        / "schema"
-        / "cube-config"
-        / "v1_4"
-        / "schema.json",  # v1 defaults to latest minor version of v1.*.
-        "//purl.org/csv-cubed/codelist-config/v1.0": APP_ROOT_DIR_PATH
-        / "schema"
-        / "codelist-config"
-        / "v1_0"
-        / "schema.json",
-        "//purl.org/csv-cubed/codelist-config/v1.1": APP_ROOT_DIR_PATH
-        / "schema"
-        / "codelist-config"
-        / "v1_1"
-        / "schema.json",
-        "//purl.org/csv-cubed/code-list-config/v1": APP_ROOT_DIR_PATH
-        / "schema"
-        / "codelist-config"
-        / "v1_1"
-        / "schema.json",  # v1 defaults to latest minor version of v1.*.
-    }
-
-    templates_dir = APP_ROOT_DIR_PATH / "readers" / "cubeconfig" / "v1_0" / "templates"
-
-    template_files = templates_dir.rglob("**/*.json*")
-
-    if not any(template_files):
-        raise ValueError(f"Couldn't find template files in {templates_dir}.")
-
-    for template_file in template_files:
-        relative_file_path = str(template_file.relative_to(templates_dir))
-        uri = (
-            "//raw.githubusercontent.com/GSS-Cogs/csvcubed/main/src/csvcubed/readers/cubeconfig/v1_0/templates/"
-            + relative_file_path
-        )
-        map_uri_to_file_path[uri] = template_file
-
-    return map_uri_to_file_path
