+++ tests/unit/utils/test_json.py
@@ -47,9 +47,10 @@ def test_load_local_when_http_request_fails():
     Ensure that a local copy of a document is returned when an http request fails in any
     instance. Most likley due to lack of internet connectivity.
     """
-
-    document = load_json_document("https://purl.org/csv-cubed/qube-config/v1.4X")
-    assert document == True
+    from csvcubed.utils.cache import session
+    with session.cache_disabled():  
+        document = load_json_document("https://purl.org/csv-cubed/qube-config/v1.4X")
+        assert document == True
 
 
 if __name__ == "__main__":
